Routing added, api class added, club views added.
A folder created named pages and the pages that contains
all of the components related to it in a independent route
like create page, edit page placed here. Pages for CRUD operations
for club model added under of the pages/club/. A blank WelcomePage.js
created for / (index root) root.
Bootstrap and Fontawesome added for react, they also supplies react
components to directly use.
Navbar created with current routing and added as a component to the
components folder.
RouterComponent is the component that handles all this routing
operations transformed to the component under the components folder.